requirements.txt
pandas
numpy
matplotlib
seaborn
psycopg2-binary
scikit-learn
xgboost
scipy
jupyter
ipykernel
